+++ model/util.c
@@ -8,7 +8,7 @@
 tw_lpid transition_select(tw_rng_stream *g,
 			  const struct transition *tr,
 			  tw_lpid limit,
-			  long int *rng_calls) {
+			  uint16_t *rng_calls) {
   tw_lpid i;
   int remaining;
   unsigned int weight;
